+++ builders/drupal-mysql.js
@@ -7,9 +7,9 @@ const LandoMysql = require('./../node_modules/@lando/mysql/builders/mysql.js');
 module.exports = {
   name: 'drupal-mysql',
   parent: '_service',
-  builder: (parent, config) => class DrupalMysql extends LandoMysql.builder(parent, LandoMysql.config) {
+  builder: parent => class DrupalMysql extends LandoMysql.builder(parent, LandoMysql.config) {
     constructor(id, options = {}) {
       super(id, options, {services: _.set({}, options.name)});
-    };
+    }
   },
 };