+++ builders/drupal10.js
@@ -38,6 +38,6 @@ module.exports = {
       }
       // Send it downstream
       super(id, options);
-    };
+    }
   },
 };