+++ builders/drupal8.js
@@ -32,6 +32,6 @@ module.exports = {
         }};
       }
       super(id, options);
-    };
+    }
   },
 };